+++ product_portfolio/tests/__init__.py
@@ -8,6 +8,8 @@
 
 from component_catalog.models import Component
 from component_catalog.models import Package
+from component_catalog.tests import make_component
+from component_catalog.tests import make_package
 from dje.tests import make_string
 from product_portfolio.models import Product
 from product_portfolio.models import ProductComponent
@@ -30,10 +32,34 @@ def make_product(dataspace, inventory=None, **data):
 
     for instance in inventory:
         if isinstance(instance, Package):
-            ProductPackage.objects.create(product=product, package=instance, dataspace=dataspace)
+            make_product_package(product, package=instance)
         if isinstance(instance, Component):
-            ProductComponent.objects.create(
-                product=product, component=instance, dataspace=dataspace
-            )
+            make_product_component(product, component=instance)
 
     return product
+
+
+def make_product_package(product, package=None):
+    dataspace = product.dataspace
+
+    if not package:
+        package = make_package(dataspace)
+
+    return ProductPackage.objects.create(
+        product=product,
+        package=package,
+        dataspace=dataspace,
+    )
+
+
+def make_product_component(product, component=None):
+    dataspace = product.dataspace
+
+    if not component:
+        component = make_component(dataspace)
+
+    return ProductComponent.objects.create(
+        product=product,
+        component=component,
+        dataspace=dataspace,
+    )